fix(cache): repair Listable build breaks, move PrefetchKeys internal
- Update the metrics test and the integration key-listing helper to the
  new Listable interface (Keys moved off V1), fixing the compile failures
  that broke the CGO/Native test and golangci-lint jobs.
- Move PrefetchKeys out of the public service SDK into the internal cache
  package so it is only available to bento's own components, per review;
  rewire aws_s3 and gcp_cloud_storage to cache.PrefetchKeys and relocate
  its tests.
